patches mainly associated with supporting gtseq q/c

+++ run_mapping_preview.sh
@@ -27,7 +27,15 @@ BCL2FASTQ_FOLDER=${WORKING_FOLDER}/../bcl2fastq/
 TAX_FOLDER=${WORKING_FOLDER}/../taxonomy_analysis
 PARAMETERS_FILE=${WORKING_FOLDER}/../../${RUN}.SampleProcessing.json
 RUN_ROOT=${WORKING_FOLDER}/..
-HPC_RESOURCE=condor
+
+if [ -z "$HPC_RESOURCE" ]; then
+   HPC_RESOURCE=local
+fi
+
+echo "HPC_RESOURCE=$HPC_RESOURCE"
+
+
+BATCHSIZE=4
 
 if [ ! -f $PARAMETERS_FILE ]; then
    echo "$PARAMETERS_FILE missing - please run get_processing_parameters.py (for help , ./get_processing_parameters.py -h)"
@@ -53,10 +61,32 @@ if [ ! -f .tardishrc ]; then
 fi
 
 
+function get_files_by_size() {
 # for each sample , the taxonomy analysis folder contains files Sample_[sample_name].list.sample.fastq.trimmed.gz 
 # these are what we need for the mapping preview  
+   lof=$1
+   tmpfile=`mktemp`
+   for sample_trimmed_file in $TAX_FOLDER/*.fastq.trimmed.gz; do
+      # get the (real) size of the file
+      byte_count=`du -bL $sample_trimmed_file | awk '{print $1}' -`
+      echo "$byte_count $sample_trimmed_file" | awk '{printf("%015d,%s\n", $1, $2);}' - >> $tmpfile
+   done
+   sort -r $tmpfile > ${tmpfile}.srt
+   awk -F, '{print $2}' ${tmpfile}.srt > $lof
+}
+
+
+# make a list of list files sorted by filesize (biggest first) so we can prioritise
+get_files_by_size ${RUN_ROOT}/lof.txt
+
+# now loop through launching BATCHSIZE at a time
 # (go back to the original bclfastq results to get the sample names)
-for sample_trimmed_file in $TAX_FOLDER/*.fastq.trimmed.gz; do
+
+sample_count=`wc ${RUN_ROOT}/lof.txt | awk '{print $1}' -`
+launched_count=0
+launched_list=`mktemp`
+set -x
+for sample_trimmed_file in `cat ${RUN_ROOT}/lof.txt`; do
    sample_name=`basename $sample_trimmed_file .list.sample.fastq.trimmed.gz`
    batonfile=${WORKING_FOLDER}/${sample_name}.bwa.baton
    if [ ! -d $WORKING_FOLDER/$sample_name ]; then
@@ -69,50 +99,55 @@ for sample_trimmed_file in $TAX_FOLDER/*.fastq.trimmed.gz; do
       if [ ! -f ${sample_trimmed_moniker}_vs_${REF_GENOME_MONIKER}.bam ]; then 
          echo "queuing $sample_trimmed_file"
          REF_GENOME_MONIKER=`basename $REF_GENOME`
-         set -x
          if [ $DRY_RUN == "yes" ]; then
             echo "*** dry run only ***"
             echo "nohup tardis.py -w -k -c 200 -hpctype $HPC_RESOURCE -d $WORKING_FOLDER/$sample_name -batonfile $batonfile bwa aln $alignment_parameters $REF_GENOME _condition_fastq_input_$sample_trimmed_file \> _condition_throughput_${sample_trimmed_moniker}_vs_${REF_GENOME_MONIKER}.sai \; bwa samse $REF_GENOME _condition_throughput_${sample_trimmed_moniker}_vs_${REF_GENOME_MONIKER}.sai _condition_fastq_input_$sample_trimmed_file  \> _condition_sam_output_$WORKING_FOLDER/${sample_trimmed_moniker}_vs_${REF_GENOME_MONIKER}  & "
          else 
             nohup tardis.py -w -k -c 200 -hpctype $HPC_RESOURCE -d $WORKING_FOLDER/$sample_name -batonfile $batonfile bwa aln $alignment_parameters $REF_GENOME _condition_fastq_input_$sample_trimmed_file \> _condition_throughput_${sample_trimmed_moniker}_vs_${REF_GENOME_MONIKER}.sai \; bwa samse $REF_GENOME _condition_throughput_${sample_trimmed_moniker}_vs_${REF_GENOME_MONIKER}.sai _condition_fastq_input_$sample_trimmed_file  \> _condition_sam_output_$WORKING_FOLDER/${sample_trimmed_moniker}_vs_${REF_GENOME_MONIKER}  & 
+            let launched_count=$launched_count+1
+            echo $sample_trimmed_file >> $launched_list
+
+
+            # if we have launched BATCHSIZE, or all of them, wait for this batch to complete
+            let modcount=${launched_count}%$BATCHSIZE
+            if [[ (  $modcount == 0 ) || (  ${launched_count} == $sample_count ) ]]; then
+               for sample_trimmed_file in `cat $launched_list`; do
+                  sample_name=`basename $sample_trimmed_file .list.sample.fastq.trimmed.gz`
+                  batonfile=${WORKING_FOLDER}/${sample_name}.bwa.baton
+                  sample_trimmed_moniker=${sample_name}.sample.fastq.trimmed
+                  REF_GENOME=`$GBS_BIN/get_processing_parameters.py --parameter_file ${PARAMETERS_FILE} --parameter_name bwa_alignment_reference  --sample $sample_name`
+                  if [  ! -z $REF_GENOME ]; then
+                     REF_GENOME_MONIKER=`basename $REF_GENOME`
+                     echo "waiting for $batonfile"
+                     while [ ! -f $batonfile ]; do  
+                        sleep 180 
+                    done
+                  fi
+                  # summarise the alignment  
+                  rm -f $batonfile
+                  if [ -f $WORKING_FOLDER/${sample_trimmed_moniker}_vs_${REF_GENOME_MONIKER}.bam ]; then
+                     bamtools stats -in $WORKING_FOLDER/${sample_trimmed_moniker}_vs_${REF_GENOME_MONIKER}.bam > $WORKING_FOLDER/${sample_trimmed_moniker}_vs_${REF_GENOME_MONIKER}.stats
+                  fi
+               done
+               launched_list=`mktemp`
+            fi
          fi
-         set +x
+      else
+         echo "skipping $sample_name (already done)"
+         # increment launched count though so we know when done
+         let launched_count=$launched_count+1
       fi
    else
       echo "skipping $sample_name (no reference specified)"
+      # increment launched count though so we know when done
+      let launched_count=$launched_count+1
    fi
 done
 
-# wait for and summarise each output file 
-for sample_trimmed_file in $TAX_FOLDER/*.fastq.trimmed.gz; do
-   sample_name=`basename $sample_trimmed_file .list.sample.fastq.trimmed.gz`
-   batonfile=${WORKING_FOLDER}/${sample_name}.bwa.baton
-   sample_trimmed_moniker=${sample_name}.sample.fastq.trimmed
-   REF_GENOME=`$GBS_BIN/get_processing_parameters.py --parameter_file ${PARAMETERS_FILE} --parameter_name bwa_alignment_reference  --sample $sample_name`
-   if [[ ( ! -z $REF_GENOME )  &&  ( $DRY_RUN == "no" ) ]]; then
-      REF_GENOME_MONIKER=`basename $REF_GENOME`
-      echo "waiting for $batonfile"
-      while [ ! -f $batonfile ]; do  
-         sleep 180 
-      done
-   fi
-   # summarise the alignment  
-   if [ $DRY_RUN == "yes" ]; then 
-      echo "
-   rm -f $batonfile
-   if [ -f $WORKING_FOLDER/${sample_trimmed_moniker}_vs_${REF_GENOME_MONIKER}.bam ]; then
-      bamtools stats -in $WORKING_FOLDER/${sample_trimmed_moniker}_vs_${REF_GENOME_MONIKER}.bam > $WORKING_FOLDER/${sample_trimmed_moniker}_vs_${REF_GENOME_MONIKER}.stats 
-   fi
-   "
-   else
-      rm -f $batonfile
-      if [ -f $WORKING_FOLDER/${sample_trimmed_moniker}_vs_${REF_GENOME_MONIKER}.bam ]; then
-         bamtools stats -in $WORKING_FOLDER/${sample_trimmed_moniker}_vs_${REF_GENOME_MONIKER}.bam > $WORKING_FOLDER/${sample_trimmed_moniker}_vs_${REF_GENOME_MONIKER}.stats
-      fi
-   fi
-done
 
 # do a grand summary  ?
+
+set +x 
 
 echo "*** run_mapping_preview.sh has completed ***"
